Rename useInvite to useAdminInvite in Admins page
